+++ src/io/segment/android/Constants.java
@@ -56,6 +56,7 @@ public class Constants {
 	public class SharedPreferences {
 		public static final String SESSION_ID_KEY = "session.id";
 		public static final String USER_ID_KEY = "user.id";
+		public static final String GROUP_ID_KEY = "group.id";
 	}
 	
 	public class Permission {