+++ tests/src/io/segment/android/test/BasicAnalyticsTest.java
@@ -19,13 +19,15 @@ public class BasicAnalyticsTest extends BaseTest {
 
 	
 	private static String userId = "android_user_" + (new Random()).nextInt(999999);
+	private static String groupId = "group_id_" + (new Random()).nextInt(999999);
 
 	
 	@Override
 	protected void setUp() {
 		super.setUp();
 		
-		Log.w("AnalyticsTest", "Analytics Test using userId: " + userId);
+		Log.w("AnalyticsTest", "Analytics Test using userId: " + 
+				userId + " and groupId: " + groupId);
 	}
 	
 	@Test
@@ -72,6 +74,49 @@ public class BasicAnalyticsTest extends BaseTest {
 	}
 	
 	@Test
+	public void testGroup() {
+
+		AnalyticsStatistics statistics = Analytics.getStatistics();
+		
+		int insertAttempts = statistics.getInsertAttempts().getCount();
+		int groupAttempts = statistics.getGroups().getCount();
+		int flushAttempts =  statistics.getFlushAttempts().getCount();
+		int successful = statistics.getSuccessful().getCount();
+
+		Traits traits = new Traits(
+			"name", "segment.io",
+			"plan", "Pro"
+		);
+		
+		Analytics.group(groupId, traits);
+		
+		Analytics.group(groupId);
+		
+		Analytics.group(new Traits(
+			"username", userId,
+			"baller", true,
+			"just_user_id", true
+		));
+		
+		Analytics.group(traits, TestCases.calendar);
+		
+		Analytics.group(traits, TestCases.calendar, new Context(
+			"providers", new EasyJSONObject(
+					"Mixpanel", true,
+					"KISSMetrics", true
+		)));
+		
+		Assert.assertEquals(groupAttempts + 5, statistics.getGroups().getCount());
+		Assert.assertEquals(insertAttempts + 5, statistics.getInsertAttempts().getCount());
+		
+		Analytics.flush(false);
+		
+		Assert.assertEquals(flushAttempts + 1, statistics.getFlushAttempts().getCount());
+		
+		Assert.assertEquals(successful + 5, statistics.getSuccessful().getCount());
+	}
+	
+	@Test
 	public void testTrack() {
 
 		AnalyticsStatistics statistics = Analytics.getStatistics();
@@ -99,6 +144,43 @@ public class BasicAnalyticsTest extends BaseTest {
 		)));
 		
 		Assert.assertEquals(trackAttempts + 5, statistics.getTracks().getCount());
+		Assert.assertEquals(insertAttempts + 5, statistics.getInsertAttempts().getCount());
+		
+		Analytics.flush(false);
+		
+		Assert.assertEquals(flushAttempts + 1, statistics.getFlushAttempts().getCount());
+		
+		Assert.assertEquals(successful + 5, statistics.getSuccessful().getCount());
+	}
+	
+	@Test
+	public void testScreen() {
+
+		AnalyticsStatistics statistics = Analytics.getStatistics();
+		
+		int insertAttempts = statistics.getInsertAttempts().getCount();
+		int screenAttempts = statistics.getScreens().getCount();
+		int flushAttempts =  statistics.getFlushAttempts().getCount();
+		int successful = statistics.getSuccessful().getCount();
+		
+		Analytics.screen("Android: Test Screen");
+		
+		Analytics.screen("Android: Another Screen Action");
+		
+		Analytics.screen("Android: First Event Screen Properties Event", new Props(
+			"Mickey Mouse", 4,
+			"Donnie", "Darko"
+		));
+		
+		Analytics.screen("Android: Screen With Calendar", new Props(),  TestCases.calendar);
+		
+		Analytics.screen("Android: Screen With Context", new Props(),  TestCases.calendar, new Context(
+			"providers", new EasyJSONObject(
+					"Mixpanel", true,
+					"KISSMetrics", true
+		)));
+		
+		Assert.assertEquals(screenAttempts + 5, statistics.getScreens().getCount());
 		Assert.assertEquals(insertAttempts + 5, statistics.getInsertAttempts().getCount());
 		
 		Analytics.flush(false);